Rename JsonUtils header
